refactor routes and extract helper functions

+++ API/routes/listRoute.js
@@ -1,41 +1,39 @@
 const express = require('express');
 const router = express.Router();
+
 const List = require('../models/listModel');
-const Task = require('../models/taskModel');
 
-const getAllLists = (res) => List.find().then((foundLists) => res.json(foundLists));
-const getAllTasks = async (res, ids = []) => {
-  const filter = ids?.length ? { _id: { $in: ids } } : {};
-  const foundTasks = await Task.find(filter);
+const { deleteTasksFromDB, getListsFromDB, getTasksFromDB } = require('../helpers/databaseHelpers');
 
-  const tasks = foundTasks.map((doc) => {
-    const task = doc.toObject();
-
-    return {
-      ...task,
-      title: task.title || task.task,
-    };
-  });
-
-  res.json(tasks);
-};
-router.route('/').get((req, res) => {
-  getAllLists(res);
+router.route('/').get(async (req, res) => {
+  try {
+    const lists = await getListsFromDB();
+    return res.json(lists);
+  } catch (err) {
+    console.error(err);
+    return res.status(500).send('Oops! Something happened!');
+  }
 });
 
 router.route('/').post((req, res) => {
   const title = req.body.title;
   const note = req.body.note;
-  const listId = req.body.listId;
 
   const newList = new List({
-    title,
     note,
-    listId,
+    tasks: [],
+    title,
   });
 
-  newList.save().then(() => {
-    getAllLists(res);
+  newList.save().then(async () => {
+    try {
+      const lists = await getListsFromDB();
+
+      return res.json(lists);
+    } catch (err) {
+      console.error(err);
+      return res.status(500).send('Oops! Something happened!');
+    }
   });
 });
 
@@ -44,11 +42,20 @@ router.route('/:listId').get(async (req, res) => {
 
   if (listId) {
     try {
-      const list = await List.findById(listId);
+      const lists = await getListsFromDB(listId);
 
-      const tasks = list.tasks;
+      if (lists?.length) {
+        const taskIds = lists?.[0]?.tasks;
 
-      return getAllTasks(res, tasks);
+        if (taskIds.length) {
+          const tasks = await getTasksFromDB(taskIds);
+          return res.json(tasks);
+        }
+
+        return res.send([]);
+      }
+
+      return res.sendStatus(400).send('No list found with that ID');
     } catch (err) {
       console.error(err);
       return res.status(500).send('Oops! Something happened!');
@@ -63,29 +70,53 @@ router.route('/:listId').post((req, res) => {
   const list = req.body.list;
   const note = req.body.note;
 
-  List.findByIdAndUpdate(listId, { list: list, note: note }) //I want this to update any changes - or just update all as if one doesn't change it's the same anyway
-    .then(() => {
-      getAllLists(res);
-      console.log('res');
+  List.findByIdAndUpdate(listId, { list: list, note: note })
+    .then(async () => {
+      try {
+        const lists = await getListsFromDB();
+        return res.json(lists);
+      } catch (err) {
+        console.error(err);
+        return res.status(500).send('Oops! Something happened!');
+      }
     })
     .catch((err) => {
       console.log(`Failed to update list ${listId}`, err);
       res.sendStatus(500);
     });
-  console.log('meow');
 });
 
-router.route('/:listId').delete((req, res) => {
+router.route('/:listId').delete(async (req, res) => {
   const listId = req.params.listId;
 
-  List.findByIdAndDelete(listId)
-    .then(() => {
-      getAllLists(res);
-    })
-    .catch((err) => {
-      console.log(`Failed deleteing task ${listId}`, err);
-      res.sendStatus(500);
-    });
+  if (listId) {
+    try {
+      const list = await List.findById(listId);
+      const tasks = list.tasks;
+
+      tasks.length && (await deleteTasksFromDB(tasks));
+
+      return List.findByIdAndDelete(listId)
+        .then(async () => {
+          try {
+            const lists = await getListsFromDB();
+            return res.json(lists);
+          } catch (err) {
+            console.error(err);
+            return res.status(500).send('Oops! Something happened!');
+          }
+        })
+        .catch((err) => {
+          console.log(`Failed deleteing task ${listId}`, err);
+          res.sendStatus(500);
+        });
+    } catch (err) {
+      console.error(err);
+      return res.status(500).send('Oops! Something happened!');
+    }
+  }
+
+  return res.status(400).send('need a list id!');
 });
 
 module.exports = router;
